+++ docs/misc/changelog.rst
@@ -28,6 +28,7 @@ Bug Fixes:
 - Allowed model trained with Python 3.7 to be loaded with Python 3.8+ without the ``custom_objects`` workaround
 - Raise an error when the same gym environment instance is passed as separate environments when creating a vectorized environment with more than one environment. (@Rocamonde)
 - Fix type annotation of ``model`` in ``evaluate_policy``
+- Fixed ``Self`` return type using ``TypeVar``
 
 Deprecations:
 ^^^^^^^^^^^^^